+++ src/sksundae/_cy_cvode.pyx
@@ -32,6 +32,7 @@ from ._cy_common import DTYPE, INT_TYPE, config  # Python precisions/config
 
 # Local python dependencies
 from .utils import RichResult
+from .cvode._precond import CVODEPrecond
 
 
 # Messages shorted from documentation online:
@@ -143,6 +144,53 @@ cdef int _jacfn_wrapper(sunrealtype t, N_Vector yy, N_Vector fy, SUNMatrix JJ,
     return 0
 
 
+cdef int _psetup_wrapper(sunrealtype t, N_Vector yy, N_Vector fy,
+                         sunbooleantype jok, sunbooleantype* jcurPtr,
+                         sunrealtype gamma, void* data) except? -1:
+    """Wraps 'psetup' by converting between N_Vector and ndarray types."""
+    
+    aux = <AuxData> data
+    psetup = aux.precond.setupfn
+
+    svec2np(yy, aux.np_yy)
+    svec2np(fy, aux.np_fy)
+
+    jnew = list((jcurPtr[0],))
+
+    if aux.with_userdata:
+        _ = psetup(t, aux.np_yy, aux.np_fy, jok, jnew, gamma, aux.userdata)
+    else:
+        _ = psetup(t, aux.np_yy, aux.np_fy, jok, jnew, gamma)
+
+    jcurPtr[0] = 1 if jnew[0] else 0
+
+    return 0
+
+
+cdef int _psolve_wrapper(sunrealtype t, N_Vector yy, N_Vector fy, N_Vector rv,
+                         N_Vector zv, sunrealtype gamma, sunrealtype delta,
+                         int lr, void* data) except? -1:
+    """Wraps 'psolve' by converting between N_Vector and ndarray types."""
+    
+    aux = <AuxData> data
+    psolve = aux.precond.solvefn
+
+    svec2np(yy, aux.np_yy)
+    svec2np(fy, aux.np_fy)
+    svec2np(rv, aux.np_rv)
+
+    if aux.with_userdata:
+        _ = psolve(t, aux.np_yy, aux.np_fy, aux.np_rv, aux.np_zv, gamma,
+                   delta, lr, aux.userdata)
+    else:
+        _ = psolve(t, aux.np_yy, aux.np_fy, aux.np_rv, aux.np_zv, gamma,
+                   delta, lr)
+
+    np2svec(aux.np_zv, zv)
+
+    return 0
+
+
 cdef void _err_handler(int line, const char* func, const char* file,
                        const char* msg, int err_code, void* err_user_data,
                        SUNContext ctx) except *:
@@ -168,6 +216,8 @@ cdef class AuxData:
     cdef np.ndarray np_fy       # right-hand-side array
     cdef np.ndarray np_ee       # events array
     cdef np.ndarray np_JJ       # Jacobian matrix
+    cdef np.ndarray np_rv       # precond rvec
+    cdef np.ndarray np_zv       # precond zvec
     cdef bint with_userdata
 
     cdef object rhsfn           # Callable
@@ -175,7 +225,8 @@ cdef class AuxData:
     cdef object eventsfn        # Callable
     cdef object jacfn           # Callable
     cdef object linsolver       # str
-    cdef object sparsity        # csc_matrix, shape(NEQ, NEQ)
+    cdef object sparsity        # csc_matrix
+    cdef object precond         # CVODEPrecond
 
     def __cinit__(self, sunindextype NEQ, object options):
         self.np_yy = np.empty(NEQ, DTYPE)
@@ -197,6 +248,14 @@ cdef class AuxData:
 
         self.linsolver = options["linsolver"]
         self.sparsity = options["sparsity"]
+
+        self.precond = options["precond"]
+        if self.precond is not None:
+            self.np_rv = np.empty(NEQ, DTYPE)
+            self.np_zv = np.empty(NEQ, DTYPE)
+        else:
+            self.np_rv = np.empty(0, DTYPE)
+            self.np_zv = np.empty(0, DTYPE)
 
 
 cdef class _cvLSSparseDQJac:
@@ -292,7 +351,7 @@ cdef class _cvLSSparseDQJac:
         self.mem = mem
         self.aux.jacfn = self
 
-        if self.aux.linsolver == 'sparse':
+        if self.aux.linsolver == "sparse":
             nnz = self.sparsity.nnz
             self.aux.np_JJ = np.zeros(nnz, DTYPE)
         else:
@@ -347,6 +406,7 @@ cdef class CVODE:
             "eventsfn": None,
             "num_events": 0,
             "jacfn": None,
+            "precond": None,
         }
 
         invalid_keys = set(options.keys()) - set(self._options.keys())
@@ -364,9 +424,19 @@ cdef class CVODE:
         self._initialized = False
 
     cdef _create_linsolver(self):
+        iterative = {"gmres", "bicgstab", "tfqmr"}
         direct = {"dense", "lapackdense", "band", "lapackband", "sparse"}
 
         linsolver = self._options["linsolver"].lower()
+
+        if "band" in linsolver:
+            uband = <int> self._options["uband"]
+            lband = <int> self._options["lband"]
+        elif linsolver in iterative:
+            maxl = <int> self._options["krylov_dim"]
+
+            precond = self._options["precond"]
+            prectype = SUN_PREC_NONE if precond is None else precond._prectype
         
         if linsolver == "dense":
             self.A = SUNDenseMatrix(self.NEQ, self.NEQ, self.ctx)
@@ -377,16 +447,10 @@ cdef class CVODE:
             self.LS = SUNLinSol_LapackDense(self.yy, self.A, self.ctx)
 
         elif linsolver == "band":
-            uband = <int> self._options["uband"]
-            lband = <int> self._options["lband"]
-
             self.A = SUNBandMatrix(self.NEQ, uband, lband, self.ctx)
             self.LS = SUNLinSol_Band(self.yy, self.A, self.ctx)
 
         elif linsolver == "lapackband":
-            uband = <int> self._options["uband"]
-            lband = <int> self._options["lband"]
-
             self.A = SUNBandMatrix(self.NEQ, uband, lband, self.ctx)
             self.LS = SUNLinSol_LapackBand(self.yy, self.A, self.ctx)
 
@@ -398,16 +462,13 @@ cdef class CVODE:
             self.LS = SUNLinSol_SuperLUMT(self.yy, self.A, nthreads, self.ctx)
 
         elif linsolver == "gmres":
-            maxl = <int> self._options["krylov_dim"]
-            self.LS = SUNLinSol_SPGMR(self.yy, SUN_PREC_NONE, maxl, self.ctx)
+            self.LS = SUNLinSol_SPGMR(self.yy, prectype, maxl, self.ctx)
 
         elif linsolver == "bicgstab":
-            maxl = <int> self._options["krylov_dim"]
-            self.LS = SUNLinSol_SPBCGS(self.yy, SUN_PREC_NONE, maxl, self.ctx)
+            self.LS = SUNLinSol_SPBCGS(self.yy, prectype, maxl, self.ctx)
 
         elif linsolver == "tfqmr":
-            maxl = <int> self._options["krylov_dim"]
-            self.LS = SUNLinSol_SPTFQMR(self.yy, SUN_PREC_NONE, maxl, self.ctx)
+            self.LS = SUNLinSol_SPTFQMR(self.yy, prectype, maxl, self.ctx)
 
         if (linsolver in direct) and (self.A is NULL):
             raise MemoryError("SUNMatrix constructor returned NULL.")
@@ -479,6 +540,12 @@ cdef class CVODE:
         if self.mem is NULL:
             raise MemoryError("CVodeCreate returned a NULL pointer for 'mem'.")
 
+        # Attach AuxData - usually done in step 16, but needs to occur here,
+        # before attaching preconditioner.
+        flag = CVodeSetUserData(self.mem, <void*> self.aux)
+        if flag < 0:
+            raise RuntimeError("CVodeSetUserData - " + CVMESSAGES[flag])
+
         # 6) Initialize CVODE solver
         flag = CVodeInit(self.mem, _rhsfn_wrapper, t0, self.yy)
         if flag < 0:
@@ -509,6 +576,19 @@ cdef class CVODE:
             flag = CVodeSetJacFn(self.mem, _jacfn_wrapper)
             if flag < 0:
                 raise RuntimeError("CVodeSetJacFn - " + LSMESSAGES[flag])
+
+        precond = self._options["precond"]
+        if precond is None:
+            pass
+        elif precond.setupfn is None:
+            flag = CVodeSetPreconditioner(self.mem, NULL, _psolve_wrapper)
+            if flag < 0:
+                raise RuntimeError("CVodeSetPrecond - " + LSMESSAGES[flag])
+        else:
+            flag = CVodeSetPreconditioner(self.mem, _psetup_wrapper,
+                                          _psolve_wrapper)
+            if flag < 0:
+                raise RuntimeError("CVodeSetPrecond - " + LSMESSAGES[flag])
 
         # 12) Create nonlinear solver object (skip, use default Newton solver)
 
@@ -542,10 +622,6 @@ cdef class CVODE:
         # 16) Set optional inputs
         SUNContext_ClearErrHandlers(self.ctx)
         SUNContext_PushErrHandler(self.ctx, _err_handler, NULL)
-
-        flag = CVodeSetUserData(self.mem, <void*> self.aux)
-        if flag < 0:
-            raise RuntimeError("CVodeSetUserData - " + CVMESSAGES[flag])
 
         cdef sunrealtype first_step = <sunrealtype> self._options["first_step"] 
         flag = CVodeSetInitStep(self.mem, first_step)
@@ -1257,3 +1333,28 @@ def _check_options(options: dict) -> None:
     if (sparsity is not None) and (jacfn is not None):
         warn("Sparse Jacobian approximation will be ignored in favor of"
              " 'jacfn'.")
+
+    # precond
+    precond = options["precond"]
+    if precond is None:
+        pass
+    elif not isinstance(precond, CVODEPrecond):
+        raise TypeError("'precond' must be type CVODEPrecond.")
+    else:
+        side = {
+            "left": SUN_PREC_LEFT,
+            "right": SUN_PREC_RIGHT,
+            "both": SUN_PREC_BOTH,
+        }
+        precond._prectype = side[precond.side]
+
+        if precond.setupfn:
+            expected = (6 + with_userdata,)
+            _ = _check_signature("precond.setupfn", precond.setupfn, expected)
+
+        expected = (8 + with_userdata,)
+        _ = _check_signature("precond.solvefn", precond.solvefn, expected)
+
+    if precond and linsolver in direct:
+        raise ValueError("'precond' is not compatitle with direct linear"
+                         f" solvers: {direct}.")